move registration entry sort functionality
